+++ docs/source/known_issues.rst
@@ -3,5 +3,3 @@ Known Issues
 
 `GitHub Issues <https://github.com/Modalities/modalities/issues>`_
 
-1. hardcoded dataset path :file:`/raid/s3/opengptx/mehdi/temp/temp_data/train_text_document.bin` in :file:`config/config.yaml`
-2. Dependency on weights&biases
